Run netperf in a dedicated core
On ARM, it is observed that netperf_maerts doesn't show the line rate
(9410). This turns out that the client was the bottleneck, which is not
we want. On the client side, the netperf process is running on a core
that also gets network irqs, which resulted in running cpu 100%. The
core is usually the last one (cpu 7)

To avoid the client being a bottleneck, run the netperf process on a
dedicated core, which is hopefully different from the one getting irqs

Signed-off-by: Jintack Lim <jintack@cs.columbia.edu>